+++ src/simulacion/estadisticas/Resultados.java
@@ -6,7 +6,7 @@ import dominio.enumeraciones.TipoModulo;
 import java.util.Map;
 
 public class Resultados {
-    public final Map<TipoModulo, Integer> tamanoPromedioCola;
+    public final Map<TipoModulo, Double> tamanoPromedioCola;
     public final double tiempoPromedioVidaConexion;
     public final int numeroConexionesDescartadas;
     public final int numeroConexionesCompletadas;
@@ -15,7 +15,7 @@ public class Resultados {
 
     public Resultados(int numeroConexionesCompletadas, int numeroConexionesDescartadas,
                       int numeroConexionesExpiradas, double tiempoPromedioVidaConexion,
-                      Map<TipoModulo, Integer> tamanoPromedioCola,
+                      Map<TipoModulo, Double> tamanoPromedioCola,
                       Map<TipoModulo, Map<TipoConsulta, Double>> tiempoPromedioPorTipoConsultas) {
         this.tamanoPromedioCola = tamanoPromedioCola;
         this.tiempoPromedioVidaConexion = tiempoPromedioVidaConexion;
